Styled mobile
Mobile now works for the footer (flexbox is fuckin’ dope!). Need to add
functionality to the form though, right now it submits to 0 places.
Will have it shoot an email to me, with the sender’s email, and fade
out that entire section, with maybe a link to social media, and a
message conveying success. Add a gif.